🐛 FIX: 버그 수정
브랜치 package-lock.json의 충돌문제 해결하였고 npm i에 문제가 있는 듯
하여 npm ci로 변경 (node_module를 완전히 제거 후 다시 설치하는 형식으로
진행)
 #5

+++ src/notification/entities/notification.entity.ts
@@ -21,7 +21,7 @@ export class Notification {
     description: string;
 
     // 알림 읽음 여부 (0 = 안 읽음, 1 = 읽음)
-    @Column({ type: 'tinyint', default: 0 })
+    @Column({ type: 'smallint', default: 0 })
     read: number;
 
     // 생성 날짜
